Implement std::string and fix the compiler bugs it exposed
Add libc/include/string with std::basic_string<char> (SSO, char_traits)
and exec_string tests on x86_64.  Along the way, fix:

- Static-member overload sets and unqualified sibling static calls; const/
  non-const member overload distinction for dependent return types.
- Copy-initialization (T b = a;) now runs the copy/move/converting
  constructor for classes with a user-declared copy/move ctor instead of a
  shallow byte copy; trivially-copyable types keep the conversion path.
- Injected-class-name in template member bodies resolves to the current
  instantiation, both as a functional-cast callee and as a type (casts,
  local declarations), so swap()/substr() target emitted symbols.
- Empty string literal emission (StringEscape produced a NULL value that
  printed as "(null)" in .asciz).
- printf %lu of values with the high bit set (signed lldiv -> unsigned /,%).
- RISC-V unsigned divide/remainder and unsigned relational branches; the
  interpreter's divu/remu.  __davecc_terminate split into eh_terminate.c so
  noexcept functions link on every backend.

Also add extern "C" linkage specifications, __cplusplus, the restrict
extension keyword, and wchar_t as an integer type-specifier.

+++ risc_v_interpreter/risc_v_interpreter.c
@@ -383,13 +383,15 @@ void RISCVInterpreterCycle(RISCVInterpreter* interpreter) {
               iregs[rd] = iregs[rs1] / iregs[rs2];
               break;
             case RV_F3(divu):
-              iregs[rd] = iregs[rs1] / iregs[rs2];
+              iregs[rd] =
+                  (int64_t)((uint64_t)iregs[rs1] / (uint64_t)iregs[rs2]);
               break;
             case RV_F3(rem):
               iregs[rd] = iregs[rs1] % iregs[rs2];
               break;
             case RV_F3(remu):
-              iregs[rd] = iregs[rs1] % iregs[rs2];
+              iregs[rd] =
+                  (int64_t)((uint64_t)iregs[rs1] % (uint64_t)iregs[rs2]);
               break;
           }
           break;
